backend/app/dependencies/mongodb.py
from pymongo import MongoClient
from dotenv import load_dotenv
import os

# Load environment variables
load_dotenv()

class MongoDB:
    def __init__(self):
        self.client = MongoClient(os.getenv("MONGO_URI"))
        self.db = self.client[os.getenv("MONGO_DB_NAME")]

    def get_collection(self, name: str):
        return self.db[name]


mongodb = MongoDB()
